+++ .settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,5 @@
 eclipse.preferences.version=1
+encoding//widget/InetChecker.py=utf-8
 encoding//widget/SwitchButton.py=utf-8
+encoding//widget/Updater.py=utf-8
 encoding/main.py=utf-8
