+++ main.py
@@ -5,200 +5,16 @@ import yaml
 from pathlib import Path
 from PyQt5 import QtCore, uic, QtWidgets
 
-import subprocess
-import threading
 import time
-import socket
 from PyQt5.Qt import QIcon
 from widget.SwitchButton import SwitchButton
-from twisted.trial.test import moduleself
+import subprocess
+from widget.Updater import Updater
+from widget.InetChecker import InetChecker
 
 USER = subprocess.check_output("logname", shell=True).rstrip().decode()
 USER_HOME_DIR = os.path.join("/home", str(USER))
 WORK_DIRECTORY = os.path.join(USER_HOME_DIR, ".life")
-
-
-class Updater(threading.Thread):
-    """ in order to provide a NONBLocking loop that 
-    periodically checks the internet connection 
-    this is done it a separate thread
-    """
-    def __init__(self, mainui):
-        threading.Thread.__init__(self)
-        self.mainui= mainui
-        self.stop = False
-
-    def run(self):
-        while self.stop == False:
-            self.update()
-            time.sleep(5)
-            
-            
-    def update(self):
-        #update life EXAM
-        line = "Updating LiFE Exam...\n"
- 
-        self.mainui.line = line
-        self.mainui.updatesignal.emit()
-    
-        cmd = "cd %s/applications/life-exam && git pull " %(WORK_DIRECTORY)
-        proc = subprocess.Popen(cmd,  shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE, bufsize=1)
-        for line in iter(proc.stderr.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        
-        for line in iter(proc.stdout.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        proc.communicate()     
-        
-        time.sleep(1)
-
-
-
-        #update life nextcloudusers
-        line = "\nUpdating LiFE Nextcloudusers...\n"
-        self.mainui.line = line
-        self.mainui.updatesignal.emit()
-        
-        cmd = "cd %s/applications/life-nextcloudusers && git pull " %(WORK_DIRECTORY)
-        proc1 = subprocess.Popen(cmd,  shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE, bufsize=1)
-        for line in iter(proc1.stderr.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        
-        for line in iter(proc1.stdout.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        proc1.communicate() 
-        
-        
-        time.sleep(1)   
-        
-        
-        
-        
-        #update life UPDATE
-        line = "\nUpdating LiFE Update...\n"
-        self.mainui.line = line
-        self.mainui.updatesignal.emit()
-        
-        cmd = "cd %s/applications/life-update && git pull " %(WORK_DIRECTORY)
-        proc2 = subprocess.Popen(cmd,  shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE, bufsize=1)
-        for line in iter(proc2.stderr.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        
-        for line in iter(proc2.stdout.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        proc2.communicate() 
-        
-        
-        time.sleep(1)   
- 
-     
-     
-        #update life FIRSTSTART
-        line = "\nUpdating LiFE Firststart...\n"
-        self.mainui.line = line
-        self.mainui.updatesignal.emit()
-        
-        cmd = "cd %s/applications/life-firststart && git pull " %(WORK_DIRECTORY)
-        proc3 = subprocess.Popen(cmd,  shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE, bufsize=1)
-        for line in iter(proc3.stderr.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        
-        for line in iter(proc3.stdout.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        proc3.communicate() 
-        
-        
-        time.sleep(1)   
-     
-        
-        #update life builder
-        line = "\nUpdating LiFE Builder...\n"
-        self.mainui.line = line
-        self.mainui.updatesignal.emit()
-        
-        cmd = "cd %s/applications/life-builder && git pull " %(WORK_DIRECTORY)
-        proc4 = subprocess.Popen(cmd,  shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE, bufsize=1)
-        for line in iter(proc4.stderr.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        
-        for line in iter(proc4.stdout.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        proc4.communicate() 
-        time.sleep(1)   
-        
-        
-        #update life kiosk
-        line = "\nUpdating LiFE Kiosk...\n"
-        self.mainui.line = line
-        self.mainui.updatesignal.emit()
-        
-        cmd = "cd %s/applications/life-kiosk && git pull " %(WORK_DIRECTORY)
-        proc5 = subprocess.Popen(cmd,  shell=True,stdout=subprocess.PIPE,stderr=subprocess.PIPE, bufsize=1)
-        for line in iter(proc5.stderr.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        
-        for line in iter(proc5.stdout.readline, b''):
-            if line:
-                self.mainui.line = line.decode()
-                self.mainui.updatesignal.emit()
-        proc5.communicate() 
-        
-        
-        time.sleep(1) 
-
-        self.mainui.finishedsignal.emit()
-        self.stop = True
-
-
-class InetChecker(threading.Thread):
-    """ in order to provide a NONBLocking loop that 
-    periodically checks the internet connection 
-    this is done it a separate thread
-    """
-    def __init__(self, mainui):
-        threading.Thread.__init__(self)
-        self.mainui = mainui
-
-    def run(self):
-        while self.mainui.check == True:
-            time.sleep(5)
-            self._checkOnline()
-
-    def _checkOnline(self):
-        try:
-            s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-            s.connect(("github.com",80))
-            s.close()
-            if self.mainui.check == True:
-                print ("online")
-                self.mainui.onsignal.emit()
-            return True
-        except:
-            print ("offline")     
-            self.mainui.offsignal.emit()
-            return False
 
 
 class MeinDialog(QtWidgets.QDialog):
@@ -231,18 +47,16 @@ class MeinDialog(QtWidgets.QDialog):
         self.check = True;
         self.line = ""
         
-        #load Config
-        self.loadConfig()
-        
         #Switch Button
         layout = self.ui.devLayout
-        layout.removeWidget(self.ui.dummySwitch)
-        
+        layout.removeWidget(self.ui.dummySwitch)        
         
         # Text, LabelOn xPos, Text, LabelOff xPos, width
-        switchbtn = SwitchButton(self, "Ja", 15, "Nein", 25, 60)
-        layout.addWidget(switchbtn)
-
+        self.switchbtn = SwitchButton(self, "Ja", 15, "Nein", 25, 60)
+        layout.addWidget(self.switchbtn)
+        
+        #load Config
+        self.loadConfig()
         
 
     def fixFilePermissions(self, folder):
@@ -278,8 +92,9 @@ class MeinDialog(QtWidgets.QDialog):
     
     def onUpdate(self): 
         self.ui.update.setEnabled(False)
+        #stop Inet Checker
         self.check = False;
-        update = Updater(self)
+        update = Updater(self, WORK_DIRECTORY)
         update.start()
     
     def uifinished(self):
@@ -292,6 +107,18 @@ class MeinDialog(QtWidgets.QDialog):
         time.sleep(1) 
         self.ui.close()
         os._exit(0)
+        
+    def log(self, msg):
+        ''' send Message to Log Box '''
+        self.line = msg
+        self.updatesignal.emit()
+        
+    def toggleDev(self):
+        if self.switchbtn.getValue():
+            line = "life-exam > [Development] Version!\n"
+        else:
+            line = "life-exam > [Stable] Version!\n"
+        self.log(line)
     
     def loadConfig(self):
         if os.path.isfile(self.__config_file): 
@@ -300,15 +127,27 @@ class MeinDialog(QtWidgets.QDialog):
         
             #for section in cfg:
                 #print(section)
-            print(cfg["development"])
+            #print(cfg["development"])
+            #print(cfg["development"]["use"])
+            useit = cfg["development"]["use"]
+            if useit==1:
+                self.switchbtn.setValue(True)
+            else:
+                self.switchbtn.setValue(False)
 
-            
+            self.branches = {}
+            self.branches["stable"] = cfg["development"]["stable_branch"]
+            self.branches["dev"] = cfg["development"]["dev_branch"]
         
     def saveConfig(self):
+        useit = 0
+        if self.switchbtn.getValue():        
+            useit = 1 
         data ={
             "development": {
-                "branch": "DEV",
-                "use": 1,
+                "stable_branch": "master",
+                "dev_branch": "DEV",
+                "use": useit,
             },
         }
         
@@ -319,8 +158,13 @@ class MeinDialog(QtWidgets.QDialog):
 
 app = QtWidgets.QApplication(sys.argv)
 dialog = MeinDialog()
-dialog.ui.show()   #show user interface
-inet = InetChecker(dialog)
-inet.start()   #start inet checking thread
+#show user interface
+dialog.ui.show()   
+
+#start inet checking thread
+#inet = InetChecker(dialog)
+#inet.start()
+
+dialog.onUpdate()   
 
 sys.exit(app.exec_())
