+++ main.ui
@@ -77,7 +77,7 @@
           <bool>true</bool>
          </property>
          <property name="indent">
-          <number>12</number>
+          <number>6</number>
          </property>
         </widget>
        </item>
@@ -128,7 +128,7 @@
              </size>
             </property>
             <property name="styleSheet">
-             <string notr="true">margin-right:6px;color:#666;</string>
+             <string notr="true">color:#666;</string>
             </property>
             <property name="text">
              <string>Fix ownership</string>
@@ -147,7 +147,7 @@
           <number>6</number>
          </property>
          <property name="rightMargin">
-          <number>6</number>
+          <number>0</number>
          </property>
          <property name="bottomMargin">
           <number>6</number>
@@ -155,7 +155,7 @@
          <item>
           <widget class="QLabel" name="label">
            <property name="styleSheet">
-            <string notr="true">margin-left: 6px;</string>
+            <string notr="true"/>
            </property>
            <property name="text">
             <string>LIFE Exam Development Version verwenden?</string>
@@ -193,6 +193,12 @@
           <item>
            <widget class="QWidget" name="widget_4" native="true">
             <layout class="QVBoxLayout" name="verticalLayout_2">
+             <property name="leftMargin">
+              <number>0</number>
+             </property>
+             <property name="rightMargin">
+              <number>0</number>
+             </property>
              <item>
               <widget class="QTextBrowser" name="info">
                <property name="styleSheet">
